backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
email-validator
python-multipart
